Refactor: move NotFound exception to repondeur
