+++ repondeur/zam_repondeur/fetch/an/amendements.py
@@ -7,10 +7,10 @@ from urllib.parse import urljoin
 
 import xmltodict
 
-from zam_aspirateur.exceptions import NotFound
 from zam_aspirateur.amendements.models import Amendement, SubDiv
 
 from zam_repondeur.fetch.division import _parse_subdiv
+from zam_repondeur.fetch.exceptions import NotFound
 from zam_repondeur.fetch.http import cached_session
